% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_conversations.R
\name{adj2edg}
\alias{adj2edg}
\title{Converts adjacency matrix to edge list}
\usage{
adj2edg(mat, symmetric = TRUE, diagonal = FALSE, text = FALSE)
}
\arguments{
\item{mat}{adjacency matrix of a graph}

\item{symmetric}{default:TRUE, the adjacency matrix is symmetric}

\item{diagonal}{default:FALSE, do not keep the diagonal values in the edge list}

\item{text}{default:FALSE, do not set the column of the edge list as character vector}
}
\value{
Return edge list of the original network as matrix
}
\description{
Converts a graph represented as adjacency matrix to edge list
}
