% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_measures.R
\name{Adj_Weighted_Jaccard}
\alias{Adj_Weighted_Jaccard}
\title{Compute the Trending Matching similarity}
\usage{
Adj_Weighted_Jaccard(input_m)
}
\arguments{
\item{input_m2}{Matrix of profiles (genes x samples)}
}
\value{
Produce the patient similarity network
}
\description{
Determine which is the similarity between patients using the TM similarity on pathway specific genes
}
