% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_measures.R
\name{Adj_Weighted_Jaccard_gene}
\alias{Adj_Weighted_Jaccard_gene}
\title{Compute the Trending Matching similarity per each gene}
\usage{
Adj_Weighted_Jaccard_gene(input_m2)
}
\arguments{
\item{input_m2}{Matrix of profiles (genes x samples)}
}
\value{
Produce a similarity matrix per gene/row of the original matrix
}
\description{
Determine which is the similarity between patients using only one gene at the time to evaluate
how much each gene is contributing to the final patient similarities
}
