% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_top_clustering_bffs_v7.R
\name{area_quadr}
\alias{area_quadr}
\title{Quadrilateral area defined by four points}
\usage{
area_quadr(A, B, C, D)
}
\arguments{
\item{A}{vector with two values defining the x and y coordinates of the first point}

\item{B}{vector with two values defining the x and y coordinates of the second point}

\item{C}{vector with two values defining the x and y coordinates of the third point}

\item{D}{vector with two values defining the x and y coordinates of the fourth point}
}
\value{
Quadrilateral area defined by four points
}
\description{
Quadrilateral area defined by four points
}
