% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_measures.R
\name{autri}
\alias{autri}
\title{Compute Trending Matching similarity}
\usage{
autri(v, w = c(1, 1.4, 2))
}
\arguments{
\item{v}{vector of two values between 0 and 1, each value is a component of the TM similarity}
}
\value{
Trending Matching similarity
}
\description{
From the two components of the similarity, it computes the overall score also called as TM in the paper
}
