% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_tools.R
\name{check_elems}
\alias{check_elems}
\title{Checks vectors elements}
\usage{
check_elems(x, y)
}
\arguments{
\item{x}{A vector}

\item{y}{A vector}
}
\value{
The result of the check. 0 not match. 1 match.
}
\description{
This function checks if two vectors are exactly the same
}
