% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_tools.R
\name{clean_pathways}
\alias{clean_pathways}
\title{Clean the pathways or feature sets provided by the user}
\usage{
clean_pathways(pathways_l, geno, size = 4, n_cores = 2)
}
\arguments{
\item{geno}{Matrix of profiles (genes x samples)}

\item{n_cores}{Number greater than 1 indicating the number of cores for parallel computing}

\item{pathwayList}{List of gene sets}

\item{dim_pathway}{Number greater than 1 indicating the minimum size of the pathways}
}
\value{
Filtered and indexes pathways/feature sets
}
\description{
Maps the feature names of the matrix of patient's profiles to the corresponding name in the feature sets/pathway
Plus, it removes all the feature sets mapping only few existing features (defined by: dim_pathway parameter)
}
