% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_tools.R
\name{column_rank}
\alias{column_rank}
\title{1-100 matrix column-wise ranking}
\usage{
column_rank(geno_p, par = T)
}
\arguments{
\item{geno_p}{Matrix of profiles (genes x samples)}

\item{par}{Boolean. Default TRUE. If FALSE diseable the parallel computing}
}
\value{
Produce a new matrix s.t. each gene has the rank of its original value in the sample
}
\description{
column-wise ranks the values and then normalizes them between 1-100
}
