% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_top_clustering_bffs_v7.R
\name{corr_BB_mat}
\alias{corr_BB_mat}
\title{Corrects the values of similarities of patients belonging to the second class}
\usage{
corr_BB_mat(mat, nAs, strong = T)
}
\arguments{
\item{mat}{adjacency matrix of a patient similarity network}

\item{nAs}{number of patients composing the second class}

\item{strong}{Default TRUE, if the second class is signature class of the PSN passed as matrix (mat)}
}
\value{
correct input PSN with adjusted values of similarities
}
\description{
Given the intra and inter similarities of one patient belonging to the second class which is signature,
it increases the intra values more the inter are low, it decreases the intra values more the inter are high.
In case, the first class is not signature it corrects in the opposite way.
}
