% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_tools.R
\name{data_preparation}
\alias{data_preparation}
\title{Controls the user-provided input data}
\usage{
data_preparation(geno, info, net, pathways_l, n_cores = 5)
}
\arguments{
\item{geno}{A numeric matrix of patient profiles: features (e.g. genes) x samples}

\item{info}{A character matrix of info about the patients: patient_ID | class (e.g. clinical information)}
}
\value{
A list with the prepared geno and info matrix
}
\description{
Prepare the user-provided input data to Simpati workflow
In case this step does not produce errors, then Simpati workflow is ready to classify and enrich the data
}
