% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_conversations.R
\name{edg2adj}
\alias{edg2adj}
\title{Converts edge list to adjacency matrix}
\usage{
edg2adj(edg1)
}
\arguments{
\item{edg1}{edge list of a graph}
}
\value{
Return djacency matrix of the original network as matrix
}
\description{
Converts a graph represented as edge list to adjacency matrix
}
