% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_measures.R
\name{eig_score_adj}
\alias{eig_score_adj}
\title{Determines how much a patient is connected to its friends against the others}
\usage{
eig_score_adj(m2, nAs)
}
\arguments{
\item{m2}{Matrix of profiles (genes x samples)}

\item{nAs}{number of elements of the first group}
}
\value{
A similarity matrix for the column objects
}
\description{
Function which computes how much a node is connected to its class, its enemies and then
it produces the adjusted eigenscore centrality of the nodes. Higher are both the measures
and higher is the score
}
