% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_enrichment.R
\name{get_cancerAtlas_associations}
\alias{get_cancerAtlas_associations}
\title{Finds "cancer" --> "gene" associations with human atlas database}
\usage{
get_cancerAtlas_associations(genes, key_names, db_cancer_ATLAS)
}
\arguments{
\item{genes}{Vector of gene names that are possible to query in disgnet database}

\item{key_names}{character vector defining the tumor of interest}

\item{db_cancer_ATLAS}{human atlas database}
}
\value{
Return matrix with the genes information about the cancer in which are involved
}
\description{
Detects "cancer" --> "gene" associations with human atlas database and return the ID of the publications
confirming the associations for each gene with the user-provided key type
}
