% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_tune.R
\name{get_consensus_classes}
\alias{get_consensus_classes}
\title{Determine the consensus class of a testing patient from a run of classification}
\usage{
get_consensus_classes(PSNs_info, vars_l, best_comb)
}
\arguments{
\item{PSNs_info}{Undefined}

\item{vars_l}{Undefined}

\item{best_comb}{Undefined}
}
\value{
Produce a new matrix s.t. each gene has the rank of its original value in the sample
}
\description{
Undefined
}
