% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_tools.R
\name{get_indxs}
\alias{get_indxs}
\title{Determines testing patient class based on the strength of its similarities}
\usage{
get_indxs(PSNs_info, k1, th_v, k_feat, comp)
}
\arguments{
\item{PSNs_info}{String character of the name of the testing profile}

\item{k1}{Matrix: Original PSN with all the profiles both training and testing}

\item{th_v}{Matrix: Clustering PSN of only training profiles}

\item{k_feat}{Simpati object: List of variables}

\item{comp}{Percentage of profiles of each class to compare with testing profile}
}
\value{
A dataframe of information for the profile class prediction, precisely:
A_strength: how much is WJ similar to A profiles similarities
B_strength: how much is WJ similar to B profiles similarities
}
\description{
Determines testing patient class based on the strength of its similarities with the training
patients that populate the two classes in comparison
}
