% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_enrichment.R
\name{get_KeysDis_associations}
\alias{get_KeysDis_associations}
\title{Finds "disase" --> "gene" associations with disgnet database}
\usage{
get_KeysDis_associations(genes, key_names, db_disgnet)
}
\arguments{
\item{genes}{Vector of gene names that are possible to query in disgnet database}

\item{db_disgnet}{disgnet database}

\item{key_type_disgnet}{character vector defining the disease of interest}
}
\value{
Return db_lit: list of matrices, each one reports the publications found supporting the association
between a gene and the disease defined with by the key words
count: percentage of disease key words associated to the genes of interest
}
\description{
Detects "disase" --> "gene" associations with disgnet database and return the ID of the publications
confirming the associations for each gene with the user-provided key type
}
