% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_enrichment.R
\name{get_new_info}
\alias{get_new_info}
\title{Produces the new info matrix from Simpati's classification results}
\usage{
get_new_info(cons_res, vars_list)
}
\arguments{
\item{cons_res}{Output of the get_perfs function}

\item{vars_list}{Variable list of the Simpati's run that has produced the cons_res oject}
}
\value{
Return new info matrix of the original patients
}
\description{
After Simpati has classified patients, this function can take Simpati's classification results
and produce the new info matrix related to the original input patients
}
