% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_enrichment.R
\name{get_normalAtlas_associations}
\alias{get_normalAtlas_associations}
\title{Finds tissues associations to the genes of interest}
\usage{
get_normalAtlas_associations(genes, key_names, db_normal_ATLAS)
}
\arguments{
\item{genes}{Vector of gene names that are possible to query in disgnet database}

\item{key_names}{character vector defining the tumor of interest}

\item{db_cancer_ATLAS}{human atlas database}
}
\value{
Return list for each gene with the user-defined tissues to which is associated
}
\description{
Detects "tissue" --> "gene" associations with human atlas database and return the ID of the publications
confirming the associations for each gene with the user-provided key type
}
