% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_enrichment.R
\name{get_pathway_enrich}
\alias{get_pathway_enrich}
\title{Wrapper: Finds diseases, tumours and tissues associations to the genes of interest}
\usage{
get_pathway_enrich(
  gene_sets_l,
  key_words,
  disease_type,
  db_disgnet,
  db_cancer_ATLAS,
  db_normal_ATLAS,
  n_cores = 2,
  records_only = TRUE
)
}
\arguments{
\item{gene_sets_l}{list of genes}

\item{key_words}{character vector defining the tumor of interest}

\item{disease_type}{character vector defining the type of disease to query in disgnet}

\item{db_disgnet}{disgnet database}

\item{db_cancer_ATLAS}{human atlas database}

\item{db_normal_ATLAS}{human atlas database}

\item{n_cores}{Default 2, number of cores to use in the parallel execution of the function}

\item{records_only}{Default TRUE, returns only the number of associations and not the information}
}
\value{
Return matrix with the genes information about the tissue in which are involved
}
\description{
Detects "disase type" --> "gene" associations, Detects "disase" --> "gene" associations,
Detects "cancer" --> "gene" associations and Detects "tissue" --> "gene" associations
}
\details{
Detects "disase type" --> "gene" associations with disgnet database and return the ID of the publications
confirming the associations for each gene with the user-provided key type

Detects "disase" --> "gene" associations with disgnet database and return the ID of the publications
confirming the associations for each gene with the user-provided key type

Detects "cancer" --> "gene" associations with human atlas database and return the ID of the publications
confirming the associations for each gene with the user-provided key type

Detects "tissue" --> "gene" associations with human atlas database and return the ID of the publications
confirming the associations for each gene with the user-provided key type
}
