% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_stats_test.R
\name{get_perfs}
\alias{get_perfs}
\title{Performs the last step of classification by predicting the final class of the testing profiles}
\usage{
get_perfs(df)
}
\arguments{
\item{df}{matrix of PSN information per testing patient profile}

\item{node_profiles_db}{data frame of profiles}
}
\value{
A list with: the testing profiles, the correct a prior classes, the predicted classes and the performances
}
\description{
Performs the consensus prediction of the testing patients and retrieves the classification performances
}
