% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_tools.R
\name{get_power}
\alias{get_power}
\title{Determines the POWER of a patient similarity network}
\usage{
get_power(m_sim_cl, Aletter, Bletter)
}
\arguments{
\item{m_sim_cl}{Similarity matrix}

\item{Aletter}{Character string. One letter to recognise the first class profiles}

\item{Bletter}{Character string. One letter to recognise the second class profiles}
}
\value{
Return the power (grade of separation between classes) of the similarity matrix
}
\description{
Computes how much the two classes separate based on their similarity
}
