% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_enrichment.R
\name{get_SemType_associations}
\alias{get_SemType_associations}
\title{Finds "disase type" --> "gene" associations with disgnet database}
\usage{
get_SemType_associations(genes, key_type_disgnet, db_disgnet)
}
\arguments{
\item{genes}{Vector of gene names that are possible to query in disgnet database}

\item{key_type_disgnet}{character string defining the disease type}

\item{db_disgnet}{disgnet database}
}
\value{
Return db_lit: list of matrices, each one reports the publications found supporting the association
between a gene and the disease type defined with by the key word
count: percentage of records supporting the gene-disease type association
}
\description{
Detects "disase type" --> "gene" associations with disgnet database and return the ID of the publications
confirming the associations for each gene with the user-provided key type
}
