% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_enrichment.R
\name{get_sign_pathways}
\alias{get_sign_pathways}
\title{Detects signature pathways related to two patient classes}
\usage{
get_sign_pathways(
  geno_p,
  pathways_l1,
  vars_list,
  th_pred = 3,
  n_test = 200,
  Pv_th = 0.05,
  n_cores = 2
)
}
\arguments{
\item{geno_p}{Matrix of patient's profiles}

\item{pathways_l1}{List of feature's sets}

\item{vars_list}{Variable list of the Simpati's run created with the set_variables function}

\item{th_pred}{Default 3, Power threshold to understand if Simpati should try to perform the Best Friend Connector algorithm}

\item{n_test}{Default 200, number of permutations to determine the p.value of significancy}

\item{Pv_th}{Default 0.05, value of threshold to assess a pathway-specific PSN as significant}

\item{n_cores}{Default 2, value which sets the number of cores to use for the parallel computation}
}
\value{
Return PSN_info_df: matrix of information related to the tested pathway specific PSNs
PSN_comp_l: list of vectorized pathway specific PSNs
outlier_df: matrix which summerizes the likelihood for each patient to be outlier of its a priori own class
}
\description{
Detects the pathways that are signature between two patient classes.
It does not depend by a step of classification. It computes a patient similarity network
per pathway and finds the PSNs which significantly separate the two classes.
It tests if a signature pathway specific PSN is signficant based on a permutation approach.
}
