% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_tools.R
\name{get_strength_sim}
\alias{get_strength_sim}
\title{Determines testing patient class based on the strength of its similarities}
\usage{
get_strength_sim(tepat, m_sim, m_sim_cl, vars_list, top_th = 0.3)
}
\arguments{
\item{tepat}{String character of the name of the testing profile}

\item{m_sim}{Matrix: Original PSN with all the profiles both training and testing}

\item{m_sim_cl}{Matrix: Clustering PSN of only training profiles}

\item{vars_list}{Simpati object: List of variables}

\item{top_th}{Percentage of profiles of each class to compare with testing profile}
}
\value{
A dataframe of information for the profile class prediction, precisely:
A_strength: how much is WJ similar to A profiles similarities
B_strength: how much is WJ similar to B profiles similarities
}
\description{
Determines testing patient class based on the strength of its similarities with the training
patients that populate the two classes in comparison
}
