% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_enrichment.R
\name{get_used_pathways}
\alias{get_used_pathways}
\title{Extracts the pathways detected as signature during a Simpati's run of patient classification}
\usage{
get_used_pathways(PSNs_info_te, vars_list)
}
\arguments{
\item{PSNs_info_te}{Output of the predict_pats function}

\item{vars_list}{Variable list of the Simpati's run that has produced the cons_res oject}
}
\value{
Return the feature list of the signature pathways used in the classification
}
\description{
After Simpati has classified patients, this function can take Simpati's classification results
and provide the pathways detected as signatures
}
