% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_tools.R
\name{h_min}
\alias{h_min}
\title{Convert minutes time value in hours}
\usage{
h_min(raw_min)
}
\arguments{
\item{raw_min}{Double value}
}
\value{
New double value formatted in hours
}
\description{
Convert minutes time value in hours
}
