% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_tools.R
\name{harm_ds}
\alias{harm_ds}
\title{Harmonizer between the info matrix and the matrix of patient's profiles}
\usage{
harm_ds(geno, info)
}
\arguments{
\item{geno}{A matrix of genetic profiles: genes x samples}

\item{info}{A matrix of info about profiles: samples and binary classes}
}
\value{
A list with the prepared geno and info matrix
}
\description{
Remove patient's profiles that do not have the clinical information
Remove patient's info that do not have a profile in the feature matrix
}
