% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_plot_net.R
\name{plot_network}
\alias{plot_network}
\title{This function creates an image that visualize a Patient Similarity Network (PSN)}
\usage{
plot_network(
  m,
  top_connections = 30,
  node_highlight = "all",
  image_type = "png",
  image_name = "plot_network"
)
}
\arguments{
\item{m}{a similarity matrix}

\item{top_connections}{an integer in [1,100], percentage of top connection to visualize}

\item{node_highlight}{a string or vector of strinvs indicating the node id(s) to highlight in the network. By default
all connections in top_connection are highlited}

\item{image_type}{a string indicating the format of the output image}

\item{image_name}{a string containing the name of the output image}

\item{label1}{a string indicating the name of the first class}

\item{label2}{a string indicating the name of the second class}
}
\value{
a list representing a qgraph object. The method saves an image with the visualization
of the network
}
\description{
This function creates an image that visualize a Patient Similarity Network (PSN)
}
