% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_pred.R
\name{predict_pats}
\alias{predict_pats}
\title{Perform the classification task to predict testing patient classes}
\usage{
predict_pats(vars_l)
}
\arguments{
\item{vars_l}{Variable list of the Simpati's run created with the set_variables function}
}
\value{
Produce a matrix s.t. for each pathway, there are information about the PSN power and the signature class
}
\description{
For each pathway, it builds a patient similarity network, it determines if it is signature due to the
training patient similarities and in case, it predicts the class of the testing patient.
}
