% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_top_clustering_bffs_v7.R
\name{reduce_net}
\alias{reduce_net}
\title{Wrapper of the BFC algorithm to plot a pathway specific patient similarity network}
\usage{
reduce_net(mat, nAs, thGroup = 0.2, thW = 0.2, A_bol = T)
}
\arguments{
\item{mat}{adjacency matrix of a patient similarity network}

\item{nAs}{number of patients composing the second class}

\item{thGroup}{Default 0.2, double value between 0 and 1 determining the percentage of best friends to keep in the end as final subclass}

\item{thW}{Default 0.2, double value between 0 and 1 determining the percentage of best friends to considers for a root (1BFS are the 60 percentage of the patients most similar to the root)}

\item{A_bol}{Default TRUE, if the first class is signature class of the PSN passed as matrix (mat)}
}
\value{
The nodes of each subclass that must be kept in the plot and the nodes that must be summerized in the legend
}
\description{
Given a patient similarity network, for each patient class, it iterates the BFC algorithm in order to
create small subclasses of similar patients.
}
