% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_RWR.R
\name{rwrProp}
\alias{rwrProp}
\title{Network based propagation based on Random walk with restart}
\usage{
rwrProp(
  net,
  geno,
  norm = "row",
  n_cores = 2,
  r = 0.8,
  stop_step = 200,
  stop_delta = 1e-06,
  keep_zero_profiles = FALSE,
  keep_no_nodes = FALSE
)
}
\arguments{
\item{net}{Gene interaction network as adjacency matrix}

\item{geno}{Numeric matrix of genetic profiles}

\item{norm}{Default row, options: row or column or laplacian, it indicates the graph normalization performed on the network}

\item{n_cores}{Default 2, An integer value greater than 0 indicating the number of cores to use for parallel computing}

\item{r}{Default 0.8, A double value lower than 1 indicating the percentage of information that a gene keeps (0.8 is 80 percentage)}

\item{stop_step}{Default 200, An integer value greater than 0 indicating the number of iterations of the propagation}

\item{stop_delta}{Default 1e-06, A double value lower than 0 indicating the threshold under which all imputed propagation values are set 0}

\item{keep_zero_profiles}{Default FALSE, Boolean: TRUE if to keep profiles without information}

\item{keep_no_nodes}{Default FALSE, Boolean: TRUE if to keep profile's rows without node in the network}
}
\value{
A list with elements describing the input data which are by the method
}
\description{
This function performs propagation of the patient's profiles passed as columns of a matrix
}
