% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_tools.R
\name{scaling_matrix}
\alias{scaling_matrix}
\title{0-1 matrix standardization}
\usage{
scaling_matrix(geno_p, min = 0, max = 1, capping = F)
}
\arguments{
\item{geno_p}{Matrix of profiles (genes x samples)}

\item{min}{Integer number to set the low limit of the range standardization. Default 0}

\item{max}{Integer number to set the high limit of the range standardization. Default 1}
}
\value{
The same input matrix with standardized values
}
\description{
0-1 matrix standardization, the propagation feature values are standardized to reflect
how much are high/low with respect all the other features in the matrix
}
