% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_tools.R
\name{set_variables}
\alias{set_variables}
\title{Creates and sets the list of variable required to run Simpati}
\usage{
set_variables(
  data_l,
  dataset_name = "simpati_ds",
  disease_type = "NAN",
  key_words = "NAN",
  n_LOO = 0.5,
  k_tuning = 10,
  th_pred = 3,
  n_cores = 2
)
}
\arguments{
\item{dataset_name}{The name of the dataset or project}

\item{disease_type}{The DISGNET semantic type of the sample condition: "Neoplastic Process" or "Disease or Syndrome"}

\item{key_words}{A character vector of one or more words related to the sample condition (e.g. KIRC: Kidney, Carcinoma)}

\item{n_LOO}{A double value lower than 1 indicating the percentage of patients to test in the LOO (e.g. 0.3 is 30 percentage)}

\item{k_tuning}{An integer value greater than 10 indicating the number of training patients to use for the tuning phase (e.g. 10)}

\item{th_pred}{An integer value greater than 0 indicating the power threshold to filter signature PSNs}

\item{n_cores}{An integer value greater than 0 indicating the number of cores to use for parallel computing}

\item{data_rdy_l}{Simpati object containing checked and prepared input data}
}
\value{
A list with elements describe the input data
}
\description{
Sets the variables based on the input data and required by Simpati to run any other function of the workflow
}
