% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_measures.R
\name{sim_WJ}
\alias{sim_WJ}
\title{Determine the first component of the Trending Matching similarity}
\usage{
sim_WJ(input_m, samples_at_columns = TRUE)
}
\arguments{
\item{input_m}{Matrix of profiles (genes x samples)}

\item{samples_at_columns}{Boolean. TRUE if samples to compute similarity are at column.}
}
\value{
A similarity matrix aka PSN with the values provided with this first component measure
}
\description{
Given the matrix of patient's profiles, it measures how much the feature information varies between patients.
In case the information is similar between two patients, then this similarity is high.
}
