% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_top_clustering_bffs_v7.R
\name{strong.comm.AA}
\alias{strong.comm.AA}
\title{Calls the BFC algorithm over the first patient class}
\usage{
strong.comm.AA(mat, nAs, thW = 0.6, thGroup = 0.6, strong = TRUE)
}
\arguments{
\item{mat}{adjacency matrix of a patient similarity network}

\item{nAs}{number of patients composing the first class}

\item{thW}{Default 0.6, double value between 0 and 1 determining the percentage of best friends to considers for a root (1BFS are the 60 percentage of the patients most similar to the root)}

\item{thGroup}{Default 0.6, double value between 0 and 1 determining the percentage of best friends to keep in the end as final subclass}

\item{strong}{Default TRUE, if the first class is signature class of the PSN passed as matrix (mat)}
}
\value{
filtered input PSN without the patients that have not been considered representatives of their own class
}
\description{
Calls the BFC algorithm over the first patient class
}
