% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_top_clustering_bffs_v1_xPred.R
\name{strong.comm.xPred3}
\alias{strong.comm.xPred3}
\title{Module of the Best Friend Connector algorithm}
\usage{
strong.comm.xPred3(m, testP, thW = 0.4, thGroup = 0.2)
}
\arguments{
\item{m}{Undefined}

\item{testP}{Undefined}

\item{thW}{Undefined}

\item{thGroup}{Undefined}
}
\value{
Undefined
}
\description{
Undefined
}
