% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_tune.R
\name{tune_performances}
\alias{tune_performances}
\title{Determine the consensus class of a testing patient from a run of classification}
\usage{
tune_performances(PSNs_info, vars_l)
}
\arguments{
\item{PSNs_info}{Undefined}

\item{vars_l}{Undefined}
}
\value{
Produce a new matrix s.t. each gene has the rank of its original value in the sample
}
\description{
Undefined
}
