% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_top_clustering_bffs_v7.R
\name{twoDist}
\alias{twoDist}
\title{Euclidean distance between two points}
\usage{
twoDist(P1, P2)
}
\arguments{
\item{P1}{vector with two values defining the x and y coordinates of the first point}

\item{P2}{vector with two values defining the x and y coordinates of the second point}
}
\value{
Euclidean distance between two points
}
\description{
Euclidean distance between two points
}
