% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_conversations.R
\name{vec2m}
\alias{vec2m}
\title{Convert graph list to adjacency matrix}
\usage{
vec2m(xL)
}
\arguments{
\item{xL}{list containing vectorized adjacency matrix, its original number of columns,
its original number of columns and node names}
}
\value{
Return adjacency matrix of the original network as matrix
}
\description{
Takes a list which contains a vector of values being the
vectorized version of an adjacency matrix, the orginal number of columns, rows and
the names of the columns. Then, this function convertes the vector to the original adjacency matrix.
}
