% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_tools.R
\name{vectorizing_matrix}
\alias{vectorizing_matrix}
\title{Vectorizes a matrix}
\usage{
vectorizing_matrix(geno_p)
}
\arguments{
\item{geno_p}{Matrix}
}
\value{
List: v = vectorized matrix content, col_names = names of the orginal columns
row_names = names of the orginal rows, n_col and n_row contain the number of original columns and rows
}
\description{
This functions takes a matrix and returns a list of its components describing its vectorize form
the values, the row and col names and its dimensions
}
