% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_tune.R
\name{wrapper_cons}
\alias{wrapper_cons}
\title{Determine the consensus class of a testing patient from a run of classification}
\usage{
wrapper_cons(PSNs_info, comb)
}
\arguments{
\item{PSNs_info}{Undefined}

\item{comb}{Undefined}
}
\value{
Produce a new matrix s.t. each gene has the rank of its original value in the sample
}
\description{
Undefined
}
