% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_wrappers.R
\name{wrapper_human_mutations}
\alias{wrapper_human_mutations}
\title{Wrapper of Simpati workflow for a human gene study}
\usage{
wrapper_human_mutations(
  geno,
  info,
  net,
  pathways_l,
  dataset_name,
  disease_type,
  key_words,
  n_cores = 5,
  n_LOO = 0.3,
  seed = 0,
  test_run = TRUE
)
}
\arguments{
\item{geno}{A numeric matrix of patient profiles: features (e.g. genes) x samples}

\item{info}{A character matrix of info about the patients: patient_ID | class (e.g. clinical information)}

\item{net}{Feature interaction network as adjacency matrix (e.g. gene-gene interactio network)}

\item{pathways_l}{List of feature sets (e.g. gene lists)}

\item{dataset_name}{The name of the dataset or project}

\item{disease_type}{The DISGNET semantic type of the sample condition: "Neoplastic Process" or "Disease or Syndrome"}

\item{key_words}{A character vector of one or more words related to the sample condition (e.g. KIRC: Kidney, Carcinoma)}

\item{n_cores}{Default 5, An integer value greater than 0 indicating the number of cores to use for parallel computing}

\item{n_LOO}{Default 0.3, A double value lower than 1 indicating the percentage of patients to test in the LOO (e.g. 0.3 is 30 percentage)}

\item{seed}{Default 0, An integer value to set the random seed generator of the run}

\item{test_run}{Default TRUE, boolean indicating if to run the function with a limited numbers of sets from pathways_l to check if Simpati works properly on the input data}
}
\value{
A list of results
}
\description{
This function performs all Simpati operations, both classification and pathway enrichment, on a study which includes
genes as features describing the human patient's profiles. This function is intended to help users which do not want to go
through all the Simpati operations manually and their study falls into the wrapper application.
}
